refactor(pipeline): refactor data pipeline
- [x] move configuration logic to a separate module;
