+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust-workspace"
-version = "0.7.20"
+version = "0.7.21"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -11,7 +11,7 @@ hyper = { version = "0.14", features = ["full"] }
 tokio = { version = "1.35.1", features = ["full"] }
 meval = "0.2"
 rand = "0.8.5"
-sysinfo = "0.30.0"
+sysinfo = "0.30.3"
 # ML Packages for Model Training
 linfa = "0.7.0"
 linfa-logistic = { version = "0.7.0", features = ["serde"] }
